Created a successful GET ISBN Request

+++ src/main/java/com/BookInfo/BookInfo/Controller/BooksController.java
@@ -21,7 +21,7 @@ public class BooksController {
     }
 
     @GetMapping("/getBookByISBN/{isbn}")
-    public Books fetchDetailsByISBN(@PathVariable int isbn){
+    public Books fetchDetailsByISBN(@PathVariable Long isbn){
         return booksService.getBookDetailsByISBN(isbn);
     }
 }
